Relocate JSON parser files to a different folder
