+++ src/client/CMakeLists.txt
@@ -12,8 +12,19 @@ add_dependencies(iota_client iota_core)
 target_sources(
   iota_client
   PRIVATE "network/http_curl.c"
-          "api/json_utils.c"
           # "api/message_builder.c"
+          "api/json_parser/common.c"
+          "api/json_parser/json_utils.c"
+          "api/json_parser/unlock_blocks.c"
+          "api/json_parser/inputs.c"
+          "api/json_parser/feat_blocks.c"
+          "api/json_parser/native_tokens.c"
+          "api/json_parser/output_alias.c"
+          "api/json_parser/output_extended.c"
+          "api/json_parser/output_foundry.c"
+          "api/json_parser/output_nft.c"
+          "api/json_parser/outputs.c"
+          "api/json_parser/unlock_conditions.c"
           "api/restful/get_node_info.c"
           "api/restful/response_error.c"
           "api/restful/get_tips.c"
@@ -26,20 +37,21 @@ target_sources(
           "api/restful/find_message.c"
           "api/restful/get_outputs_id.c"
           "api/restful/get_transaction_included_message.c"
-          "api/restful/models/json_parser_common.c"
-          "api/restful/models/json_unlock_blocks.c"
-          "api/restful/models/inputs/json_inputs.c"
-          "api/restful/models/outputs/json_feat_blocks.c"
-          "api/restful/models/outputs/json_native_tokens.c"
-          "api/restful/models/outputs/json_output_alias.c"
-          "api/restful/models/outputs/json_output_extended.c"
-          "api/restful/models/outputs/json_output_foundry.c"
-          "api/restful/models/outputs/json_output_nft.c"
-          "api/restful/models/outputs/json_outputs.c"
-          "api/restful/models/outputs/json_unlock_conditions.c"
   PUBLIC "network/http.h"
-         "api/json_utils.h"
-         "api/json_keys.h"
+         "client_service.h"
+         "api/json_parser/json_utils.h"
+         "api/json_parser/json_keys.h"
+         "api/json_parser/common.h"
+         "api/json_parser/unlock_blocks.h"
+         "api/json_parser/inputs.h"
+         "api/json_parser/feat_blocks.h"
+         "api/json_parser/native_tokens.h"
+         "api/json_parser/output_alias.h"
+         "api/json_parser/output_extended.h"
+         "api/json_parser/output_foundry.h"
+         "api/json_parser/output_nft.h"
+         "api/json_parser/outputs.h"
+         "api/json_parser/unlock_conditions.h"
          # "api/message_builder.h"
          "api/restful/get_node_info.h"
          "api/restful/response_error.h"
@@ -52,19 +64,7 @@ target_sources(
          "api/restful/send_message.h"
          "api/restful/find_message.h"
          "api/restful/get_outputs_id.h"
-         "client_service.h"
          "api/restful/get_transaction_included_message.h"
-         "api/restful/models/json_parser_common.h"
-         "api/restful/models/json_unlock_blocks.h"
-         "api/restful/models/inputs/json_inputs.h"
-         "api/restful/models/outputs/json_feat_blocks.h"
-         "api/restful/models/outputs/json_native_tokens.h"
-         "api/restful/models/outputs/json_output_alias.h"
-         "api/restful/models/outputs/json_output_extended.h"
-         "api/restful/models/outputs/json_output_foundry.h"
-         "api/restful/models/outputs/json_output_nft.h"
-         "api/restful/models/outputs/json_outputs.h"
-         "api/restful/models/outputs/json_unlock_conditions.h"
          )
 
 if(MQTT_CLIENT_ENABLE)
