+++ src/client/api/json_parser/common.h
@@ -1,10 +1,10 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef __CLIENT_API_RESTFUL_MODELS_JSON_PARSER_COMMON_H__
-#define __CLIENT_API_RESTFUL_MODELS_JSON_PARSER_COMMON_H__
+#ifndef __CLIENT_API_JSON_PARSER_COMMON_H__
+#define __CLIENT_API_JSON_PARSER_COMMON_H__
 
-#include "client/api/json_utils.h"
+#include "client/api/json_parser/json_utils.h"
 #include "core/address.h"
 
 #ifdef __cplusplus
@@ -24,4 +24,4 @@ int json_parser_common_address_deserialize(cJSON *json_obj, address_t *address);
 }
 #endif
 
-#endif  // __CLIENT_API_RESTFUL_MODELS_JSON_PARSER_COMMON_H__
+#endif  // __CLIENT_API_JSON_PARSER_COMMON_H__
