+++ src/client/api/json_parser/feat_blocks.h
@@ -1,10 +1,10 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef __CLIENT_API_RESTFUL_MODELS_OUTPUTS_JSON_FEAT_BLOCKS_H__
-#define __CLIENT_API_RESTFUL_MODELS_OUTPUTS_JSON_FEAT_BLOCKS_H__
+#ifndef __CLIENT_API_JSON_PARSER_FEAT_BLOCKS_H__
+#define __CLIENT_API_JSON_PARSER_FEAT_BLOCKS_H__
 
-#include "client/api/json_utils.h"
+#include "client/api/json_parser/json_utils.h"
 #include "core/models/outputs/feat_blocks.h"
 
 #ifdef __cplusplus
@@ -24,4 +24,4 @@ int json_feat_blocks_deserialize(cJSON *output_obj, feat_blk_list_t *feat_blocks
 }
 #endif
 
-#endif  // __CLIENT_API_RESTFUL_MODELS_OUTPUTS_JSON_FEAT_BLOCKS_H__
+#endif  // __CLIENT_API_JSON_PARSER_FEAT_BLOCKS_H__
