+++ src/client/api/json_parser/json_keys.h
@@ -1,8 +1,8 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef __CLIENT_API_JSON_KEYS_H__
-#define __CLIENT_API_JSON_KEYS_H__
+#ifndef __CLIENT_API_JSON_PARSER_JSON_KEYS_H__
+#define __CLIENT_API_JSON_PARSER_JSON_KEYS_H__
 
 static char const* const JSON_KEY_ADDR = "address";
 static char const* const JSON_KEY_ADDR_TYPE = "addressType";
