+++ src/client/api/json_parser/native_tokens.h
@@ -1,10 +1,10 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef __CLIENT_API_RESTFUL_MODELS_OUTPUTS_JSON_NATIVE_TOKENS_H__
-#define __CLIENT_API_RESTFUL_MODELS_OUTPUTS_JSON_NATIVE_TOKENS_H__
+#ifndef __CLIENT_API_JSON_PARSER_NATIVE_TOKENS_H__
+#define __CLIENT_API_JSON_PARSER_NATIVE_TOKENS_H__
 
-#include "client/api/json_utils.h"
+#include "client/api/json_parser/json_utils.h"
 #include "core/models/outputs/native_tokens.h"
 
 #ifdef __cplusplus
@@ -24,4 +24,4 @@ int json_native_tokens_deserialize(cJSON *output_obj, native_tokens_t *native_to
 }
 #endif
 
-#endif  // __CLIENT_API_RESTFUL_MODELS_OUTPUTS_JSON_NATIVE_TOKENS_H__
+#endif  // __CLIENT_API_JSON_PARSER_NATIVE_TOKENS_H__
