+++ src/client/api/json_parser/output_alias.c
@@ -1,9 +1,9 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#include "client/api/restful/models/outputs/json_output_nft.h"
+#include "client/api/json_parser/output_alias.h"
 
-int json_output_nft_deserialize(cJSON *output_obj, transaction_essence_t *essence) {
+int json_output_alias_deserialize(cJSON *output_obj, transaction_essence_t *essence) {
   if (output_obj == NULL || essence == NULL) {
     printf("[%s:%d]: Invalid parameters\n", __func__, __LINE__);
     return -1;
