+++ src/client/api/restful/get_outputs_id.c
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "client/api/restful/get_outputs_id.h"
-#include "client/api/json_utils.h"
+#include "client/api/json_parser/json_utils.h"
 #include "client/network/http.h"
 #include "core/utils/iota_str.h"
