+++ src/client/api/restful/get_tips.c
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 
-#include "client/api/json_utils.h"
+#include "client/api/json_parser/json_utils.h"
 #include "client/api/restful/get_tips.h"
 #include "core/utils/iota_str.h"
