+++ src/client/api/restful/send_message.c
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 
-#include "client/api/json_utils.h"
+#include "client/api/json_parser/json_utils.h"
 #include "client/api/message_builder.h"
 #include "client/api/restful/get_tips.h"
 #include "client/api/restful/send_message.h"
